refactor(cli): split main.rs into command modules (<1000 lines)
main.rs was 2006 lines. Extract the largest pieces into commands/ submodules,
keeping each command's clap args colocated with its handler:

- commands/clawhub.rs: ClawHub* clap types + the `clawhub` handler.
- commands/swarm.rs: SwarmCommands + the `swarm` handler.
- commands/gateway_client.rs: the gateway WebSocket client helpers
  (send_gateway_reload, send_command_via_gateway, run_local_command) and the
  headless `ask` command (AskArgs + handle_ask).

main.rs (785 lines) keeps the top-level clap tree and the dispatch match,
which now delegates these commands to their modules. No behavior change.
