+++ crates/rustyclaw-cli/src/commands/mod.rs
@@ -2,12 +2,15 @@
 //!
 //! Each submodule handles a specific command group (gateway, skills, etc.)
 
+pub mod clawhub;
 pub mod config;
 pub mod gateway;
+pub mod gateway_client;
 pub mod import;
 pub mod refresh_token;
 pub mod shared;
 pub mod status;
+pub mod swarm;
 
 // Re-export handlers for use in main.rs
 pub(crate) use config::{config_get, config_set, config_unset};
